+++ src/components/ChainPage/index.js
@@ -2,7 +2,7 @@ import React, { useMemo, useState } from 'react'
 import dynamic from 'next/dynamic'
 import styled from 'styled-components'
 import { RowFixed } from '../Row'
-import Search from '../Search/New'
+import Search from '../Search/OpenSearch'
 import { ProtocolsTable, Panel, BreakpointPanels, BreakpointPanel, PanelHiddenMobile, ChartAndValuesWrapper } from '..'
 import Filters from '../Filters'
 import { useDarkModeManager, useGetExtraTvlEnabled } from 'contexts/LocalStorage'
