+++ src/components/NFTDashboardPage/index.js
@@ -4,7 +4,7 @@ import { transparentize } from 'polished'
 import { useDisplayUsdManager, useHideLastDayManager } from '../../contexts/LocalStorage'
 import Filters from '../Filters'
 import { CheckMarks } from '../SettingsModal'
-import { NFTsSearch } from '../Search/New'
+import { NFTsSearch } from '../Search/OpenSearch'
 import NFTCollectionList from '../NFTCollectionList'
 import { formattedNum } from '../../utils'
 import { chainCoingeckoIds, chainMarketplaceMappings } from '../../constants/chainTokens'
