+++ src/containers/ChainsContainer/index.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react'
 import styled from 'styled-components'
 import { ButtonDark } from 'components/ButtonStyled'
-import Search from 'components/Search/New'
+import Search from 'components/Search/OpenSearch'
 import { ChainPieChart, ChainDominanceChart } from 'components/Charts'
 import { columnsToShow, FullTable } from 'components/Table'
 import { toNiceCsvDate, getRandomColor, download } from 'utils'
