+++ src/containers/ProtocolContainer/index.tsx
@@ -10,7 +10,7 @@ import { useCalcSingleExtraTvl } from '../../hooks/data'
 import { useScrollToTop } from 'hooks'
 import { capitalizeFirstLetter, formattedNum, getBlockExplorer, toK } from 'utils'
 import SEO from 'components/SEO'
-import Search from 'components/Search/New'
+import Search from 'components/Search/OpenSearch'
 import Layout from 'layout'
 import { ArrowUpRight, DownloadCloud } from 'react-feather'
 import AuditInfo from 'components/AuditInfo'
