+++ src/pages/forks.js
@@ -6,7 +6,7 @@ import { useCalcGroupExtraTvlsByDay, useCalcStakePool2Tvl } from 'hooks/data'
 import { getRandomColor, toK } from 'utils'
 import { CustomLink } from 'components/Link'
 import styled from 'styled-components'
-import Search from 'components/Search/New'
+import Search from 'components/Search/OpenSearch'
 import { Header } from 'Theme'
 import { ChainDominanceChart, ChainPieChart } from 'components/Charts'
 import Filters, { FiltersWrapper } from 'components/Filters'
