+++ src/pages/oracles.js
@@ -6,7 +6,7 @@ import { Box } from 'rebass'
 import { CustomLink } from 'components/Link'
 import { getRandomColor, toK } from 'utils'
 import { useCalcGroupExtraTvlsByDay } from 'hooks/data'
-import Search from 'components/Search/New'
+import Search from 'components/Search/OpenSearch'
 import { Header } from 'Theme'
 import { ChainDominanceChart, ChainPieChart } from 'components/Charts'
 import Filters, { FiltersWrapper } from 'components/Filters'
